+++ src/apps/maple/maple.module.ts
@@ -4,6 +4,7 @@ import { AbstractApp } from '~app/app.dynamic-module';
 
 import { MaplePoolDefinitionResolver } from './common/maple.pool.definition-resolver';
 import { MapleContractFactory } from './contracts';
+import { EthereumMaplePendingWithdrawalContractPositionFetcher } from './ethereum/maple.pending-withdrawal.contract-position-fetcher';
 import { EthereumMaplePoolTokenFetcher } from './ethereum/maple.pool.token-fetcher';
 import { EthereumMapleXMplTokenFetcher } from './ethereum/maple.x-mpl.token-fetcher';
 
@@ -13,6 +14,7 @@ import { EthereumMapleXMplTokenFetcher } from './ethereum/maple.x-mpl.token-fetc
     MaplePoolDefinitionResolver,
     EthereumMaplePoolTokenFetcher,
     EthereumMapleXMplTokenFetcher,
+    EthereumMaplePendingWithdrawalContractPositionFetcher,
   ],
 })
 export class MapleAppModule extends AbstractApp() {}