+++ src/apps/origin-dollar-governance/ethereum/origin-dollar-governance.rewards.contract-position-fetcher.ts
@@ -3,13 +3,11 @@ import { BigNumberish } from 'ethers';
 
 import { APP_TOOLKIT, IAppToolkit } from '~app-toolkit/app-toolkit.interface';
 import { PositionTemplate } from '~app-toolkit/decorators/position-template.decorator';
-import { getLabelFromToken } from '~app-toolkit/helpers/presentation/image.present';
 import { DefaultDataProps } from '~position/display.interface';
 import { MetaType } from '~position/position.interface';
 import { ContractPositionTemplatePositionFetcher } from '~position/template/contract-position.template.position-fetcher';
 import {
   DefaultContractPositionDefinition,
-  GetDisplayPropsParams,
   GetTokenBalancesParams,
   UnderlyingTokenDefinition,
 } from '~position/template/contract-position.template.types';
@@ -17,8 +15,8 @@ import {
 import { OriginDollarContractFactory, Veogv } from '../contracts';
 
 @PositionTemplate()
-export class EthereumOriginDollarRewardsContractPositionFetcher extends ContractPositionTemplatePositionFetcher<Veogv> {
-  groupLabel = 'Reward';
+export class EthereumOriginDollarGovernanceRewardsContractPositionFetcher extends ContractPositionTemplatePositionFetcher<Veogv> {
+  groupLabel = 'Staking Reward';
 
   constructor(
     @Inject(APP_TOOLKIT) protected readonly appToolkit: IAppToolkit,
@@ -45,9 +43,8 @@ export class EthereumOriginDollarRewardsContractPositionFetcher extends Contract
     ];
   }
 
-  async getLabel({ contractPosition }: GetDisplayPropsParams<Veogv>) {
-    const rewardToken = contractPosition.tokens[0];
-    return `${getLabelFromToken(rewardToken)} Staking Rewards`;
+  async getLabel() {
+    return `Origin Dollar Governance Staking Rewards`;
   }
 
   async getTokenBalancesPerPosition({